requirements.txt
yfinance
pandas
numpy
gymnasium
stable-baselines3[extra]
PyPortfolioOpt
pyfolio-reloaded
matplotlib
seaborn
scipy
